+++ test/teslamate/vehicles/vehicle/suspend_test.exs
@@ -38,7 +38,7 @@ defmodule TeslaMate.Vehicles.Vehicle.SuspendTest do
     assert DateTime.diff(s0, s1, :nanosecond) < 0
     assert_receive {:insert_position, ^car, %{}}
 
-    assert_receive {:start_state, ^car, :asleep, []}
+    assert_receive {:start_state, ^car, :asleep, [date: _]}
     assert_receive {:"$websockex_cast", :disconnect}
     assert_receive {:pubsub, {:broadcast, _, _, %Summary{state: :asleep, since: s2}}}
     assert DateTime.diff(s1, s2, :nanosecond) < 0
@@ -426,7 +426,7 @@ defmodule TeslaMate.Vehicles.Vehicle.SuspendTest do
 
     assert_receive {:pubsub, {:broadcast, _, _, %Summary{state: :suspended}}}
     assert_receive {:insert_position, ^car, %{}}
-    assert_receive {:start_state, ^car, :asleep, []}, 50
+    assert_receive {:start_state, ^car, :asleep, [date: _]}, 50
     assert_receive {:"$websockex_cast", :disconnect}
     assert_receive {:pubsub, {:broadcast, _, _, %Summary{state: :asleep}}}